requirements.txt
Django==2.2.9
mysqlclient==1.4.6
python-decouple==3.3
Pillow==6.2.1
django-taggit==1.2.0